requirements.txt
flask
gunicorn
gevent
alembic
pydantic
Flask-Cors
psycopg2
psycogreen
requests
json-logging
python-dotenv
Flask-pydantic
sqlalchemy
pyjwt[crypto]
flask-request-id-header
